test: demo - ⭕ GestureDemoListener
